[Feat]Warrior basic attack + player HP

+++ Assets/Scripts/Units/IDamageable.cs
@@ -1,4 +1,4 @@
 public interface IDamageable
 {
-    public void GetDamage(int damage);
+    public void GetDamage(float damage);
 }
